added a bit more functionality
we need a way to display the contents of the hashmap in the gui -- preferably in a new window that the user can close when they're done with it. i was thinking we could use a JTable, which would serve as one of the four new features we need to include and allow the data to be easily and neatly displayed, but it would also require that we put the contents of a hashmap into an array, which would a) be a pain and b) defeat the purpose of using the hashmap.