Added the beginnings of the flashcard function
It's quite incomplete -- we still need to add the mouseListener or button that will add the actual flashcard functionality -- but I figured I'd push this before I go to bed. Also, I moved the displayWindow stuff out of the displayButton action listener (since putting it there is seen as bad form) and now it doesn't quite work.